Do not show annotation modal when clicking annotationless line segment
