+++ src/utils/Optional.ts
@@ -1,0 +1,4 @@
+/**
+ * Make a property optional
+ */
+export type Optional<T, K extends keyof T> = Pick<Partial<T>, K> & Omit<T, K>;
